extract commands from index.js
