+++ index.js
@@ -3,10 +3,8 @@
 var bodyParser = require('body-parser')
 var fs = require('fs')
 var path = require('path')
-var spawn = require('child_process').spawn
-var RSVP = require('rsvp')
 var request = require('sync-request')
-var chalk = require('chalk')
+var mkdirpSync = require('./lib/utils').mkdirpSync
 
 function uploadToImgur (image, reportPath) {
   var imgurClientID = 'e39f00905b80937'
@@ -41,29 +39,6 @@ function uploadToExpress (url, image, name) {
   return response
 }
 
-function runCommand (command, args, ignoreStdError) {
-  return new RSVP.Promise(function (resolve, reject) {
-    var child = spawn(command, args, {
-      env: process.env
-    })
-    child.stdout.on('data', function (data) {
-      console.log(data.toString())
-    })
-    child.stderr.on('data', function (data) {
-      console.error(chalk.bold.red(data.toString()))
-    })
-
-    child.on('exit', function (code) {
-      console.log('Exit with code ' + code)
-      if (code === 0 || ignoreStdError) {
-        resolve()
-      } else {
-        reject()
-      }
-    })
-  })
-}
-
 function compareVersions (installed, required) {
   if (required === undefined) {
     return true
@@ -84,21 +59,6 @@ function compareVersions (installed, required) {
     return true
   } else {
     return installed === required
-  }
-}
-
-function mkdirSync (path) {
-  try {
-    fs.mkdirSync(path)
-  } catch (e) {
-    if (e.code !== 'EEXIST') throw e
-  }
-}
-
-function mkdirpSync (dirpath) {
-  let parts = dirpath.split(path.sep)
-  for (let i = 1; i <= parts.length; i++) {
-    mkdirSync(path.join.apply(null, parts.slice(0, i)))
   }
 }
 
@@ -205,118 +165,6 @@ function getImage (req, res, options) {
   }
 }
 
-function buildReport (params) {
-  return new Promise(function (resolve, reject) {
-    console.log('Sending to github')
-    var markdownReport = JSON.parse(fs.readFileSync('visual-acceptance-report/report.json'))
-    var markdownBody = '# Visual Acceptance tests:\n'
-    if (markdownReport.new !== '## New\n') {
-      markdownBody += markdownReport.new + '\n'
-    }
-    if (markdownReport.changed !== '## Changed\n') {
-      markdownBody += markdownReport.changed
-    }
-    if (markdownReport.changed === '## Changed\n' && markdownReport.new === '## New\n') {
-      markdownBody += '### No changes\n'
-    }
-    try {
-      if (process.env.REPORT_MARKDOWN_PATH) {
-        fs.writeFileSync(process.env.REPORT_MARKDOWN_PATH, markdownBody)
-      }
-    } catch (e) {
-      console.log(e)
-    }
-    var githubApiPostOptions = {
-      'headers': {
-        'user-agent': 'visual-acceptance',
-        'Authorization': 'token ' + process.env.VISUAL_ACCEPTANCE_TOKEN
-      },
-      'json': {
-        'body': markdownBody
-      }
-    }
-
-    var githubApiGetOptions = {
-      'headers': {
-        'user-agent': 'visual-acceptance',
-        'Authorization': 'token ' + process.env.VISUAL_ACCEPTANCE_TOKEN
-      }
-    }
-    var url = 'https://api.github.com/repos/' + process.env.TRAVIS_REPO_SLUG + '/issues/' +
-     process.env.TRAVIS_PULL_REQUEST + '/comments'
-    var response = request('GET', url, githubApiGetOptions)
-    var bodyJSON = JSON.parse(response.getBody().toString())
-    for (var i = 0; i < bodyJSON.length; i++) {
-      if (bodyJSON[i].body.indexOf('# Visual Acceptance tests:') > -1) {
-        url = bodyJSON[i].url
-        break
-      }
-    }
-    response = request('POST', url, githubApiPostOptions)
-    resolve(response)
-  })
-}
-
-function buildTeamcityBitbucketReport (params, options, prNumber) {
-  return new Promise(function (resolve, reject) {
-    console.log('Sending to bitbucket')
-    var markdownReport = JSON.parse(fs.readFileSync('visual-acceptance-report/report.json'))
-    var markdownBody = '# Visual Acceptance tests:\n'
-    if (markdownReport.new !== '## New\n') {
-      markdownBody += markdownReport.new + '\n'
-    }
-    if (markdownReport.changed !== '## Changed\n') {
-      markdownBody += markdownReport.changed
-    }
-    if (markdownReport.changed === '## Changed\n' && markdownReport.new === '## New\n') {
-      markdownBody += '### No changes\n'
-    }
-    try {
-      if (process.env.REPORT_MARKDOWN_PATH) {
-        fs.writeFileSync(process.env.REPORT_MARKDOWN_PATH, markdownBody)
-      }
-    } catch (e) {
-      console.log(e)
-    }
-    var githubApiPostOptions = {
-      'headers': {
-        'user-agent': 'visual-acceptance',
-        'Authorization': 'Basic ' + new Buffer(options.user + ':' + options.password, 'ascii').toString('base64')
-      },
-      'json': {
-        'text': markdownBody
-      }
-    }
-    var githubApiGetOptions = {
-      'headers': {
-        'user-agent': 'visual-acceptance',
-        'Authorization': 'Basic ' + new Buffer(options.user + ':' + options.password, 'ascii').toString('base64')
-      }
-    }
-    var url = 'http://' + options.domain + '/rest/api/1.0/projects/' + options.project + '/repos/' +
-     options.repo + '/pull-requests/' + prNumber + '/comments'
-    var urlGet = 'http://' + options.domain + '/rest/api/1.0/projects/' + options.project + '/repos/' +
-     options.repo + '/pull-requests/' + prNumber + '/activities'
-    var response = request('GET', urlGet, githubApiGetOptions)
-    var bodyJSON = JSON.parse(response.getBody().toString())
-    var existingComment = false
-    for (var i = 0; i < bodyJSON.values.length; i++) {
-      if (bodyJSON.values[i].action === 'COMMENTED' &&
-        bodyJSON.values[i].comment.text.indexOf('# Visual Acceptance tests:') > -1) {
-        url += '/' + bodyJSON.values[i].comment.id
-        githubApiPostOptions.json.version = bodyJSON.values[i].comment.version
-        existingComment = true
-        break
-      }
-    }
-    if (existingComment) {
-      response = request('PUT', url, githubApiPostOptions)
-    } else {
-      response = request('POST', url, githubApiPostOptions)
-    }
-    resolve(response)
-  })
-}
 module.exports = {
   name: 'ember-cli-visual-acceptance',
   included: function (app) {
@@ -408,245 +256,6 @@ module.exports = {
   },
 
   includedCommands: function () {
-    return {
-      'build-report': {
-        name: 'build-report',
-        aliases: ['br'],
-        description: 'Create report',
-        works: 'insideProject',
-        availableOptions: [{
-          name: 'report-directory',
-          type: String,
-          default: 'visual-acceptance-report',
-          description: 'Create Report off visual acceptance tests'
-        }],
-        run: function (options, rawArgs) {
-          var root = this.project.root
-
-          function deleteFolderRecursive (path) {
-            if (fs.existsSync(path)) {
-              fs.readdirSync(path).forEach(function (file, index) {
-                var curPath = path + '/' + file
-                if (fs.lstatSync(curPath).isDirectory()) { // recurse
-                  deleteFolderRecursive(curPath)
-                } else { // delete file
-                  fs.unlinkSync(curPath)
-                }
-              })
-              fs.rmdirSync(path)
-            }
-          }
-
-          deleteFolderRecursive(path.join(root, options.reportDirectory))
-          mkdirpSync(options.reportDirectory)
-          var reportPath = options.reportDirectory + '/' + 'report.html'
-          var markdownPath = options.reportDirectory + '/' + 'report.md'
-          var jsonPath = options.reportDirectory + '/' + 'report.json'
-          var markdownReport = {
-            new: '## New\n',
-            changed: '## Changed\n'
-          }
-          fs.writeFileSync(jsonPath, JSON.stringify(markdownReport))
-          process.env.PR_API = options.prApiUrl
-          process.env.REPORT_PATH = reportPath
-          process.env.REPORT_MARKDOWN_PATH = markdownPath
-          process.env.REPORT_JSON_PATH = jsonPath
-          return runCommand('ember', ['test'])
-        }
-      },
-      'new-baseline': {
-        name: 'new-baseline',
-        aliases: ['nb'],
-        description: 'Create new baseline',
-        works: 'insideProject',
-        availableOptions: [{
-          name: 'image-directory',
-          type: String,
-          default: 'visual-acceptance',
-          description: 'The ember-cli-visual-acceptance directory where images are save'
-        }, {
-          name: 'build-report',
-          type: Boolean,
-          default: false,
-          description: 'Wheter or not to build a report'
-        }],
-        run: function (options, rawArgs) {
-          var root = this.project.root
-
-          function deleteFolderRecursive (path) {
-            if (fs.existsSync(path)) {
-              fs.readdirSync(path).forEach(function (file, index) {
-                var curPath = path + '/' + file
-                if (fs.lstatSync(curPath).isDirectory()) { // recurse
-                  deleteFolderRecursive(curPath)
-                } else { // delete file
-                  fs.unlinkSync(curPath)
-                }
-              })
-              fs.rmdirSync(path)
-            }
-          }
-
-          deleteFolderRecursive(path.join(root, options.imageDirectory))
-          if (options.buildReport) {
-            return runCommand('ember', ['br'])
-          } else {
-            return runCommand('ember', ['test'])
-          }
-        }
-      },
-      'travis-visual-acceptance': {
-        name: 'travis-visual-acceptance',
-        aliases: ['tva'],
-        description: 'Run visual-acceptance based off Travis message',
-        works: 'insideProject',
-        availableOptions: [{
-          name: 'image-directory',
-          type: String,
-          default: 'visual-acceptance',
-          description: 'The ember-cli-visual-acceptance directory where images are save'
-        }, {
-          name: 'branch',
-          type: String,
-          default: 'master',
-          description: 'branch to push to'
-        }, {
-          name: 'push-var',
-          type: String,
-          default: 'true',
-          description: 'Option that compares to pushOn to determine if it\'s okay to push'
-        },
-        {
-          name: 'push-on',
-          type: String,
-          default: 'true',
-          description: 'Option that compares to pushVar to determine if it\'s okay to push'
-        }],
-        run: function (options, rawArgs) {
-          if (!process.env.RO_GH_TOKEN || !process.env.TRAVIS_REPO_SLUG || !process.env.VISUAL_ACCEPTANCE_TOKEN) {
-            // eslint-disable-next-line max-len
-            console.log('No github token found (RO_GH_TOKEN), Travis found, or visual acceptance token (VISUAL_ACCEPTANCE_TOKEN) found. Just running ember test')
-            return runCommand('ember', ['test'])
-          }
-          process.env.NO_ASSERT = true
-
-          var prNumber = process.env.TRAVIS_PULL_REQUEST
-          if (prNumber !== false && prNumber !== 'false' && process.env.VISUAL_ACCEPTANCE_TOKEN) {
-            return runCommand('ember', ['br']).then(buildReport, function (params) {
-              return buildReport(params).then(function (params) {
-                throw new Error('Exit 1')
-              })
-            })
-          } else if (prNumber === false || prNumber === 'false') {
-            return runCommand('ember', ['new-baseline', '--image-directory=' +
-             options.imageDirectory]).then(function (params) {
-               return runCommand('git', ['add', '-f', '-A', options.imageDirectory], true).then(function (params) {
-                 return runCommand('git', ['commit', '-m',
-                  'Adding new baseline images [ci skip]'], true).then(function (params) {
-                    if (options.pushVar === options.pushOn) {
-                      return runCommand('git', ['push', 'origin', 'HEAD:' + options.branch])
-                    }
-                  })
-               })
-             })
-          }
-        }
-      },
-      'teamcity-bitbucket-visual-acceptance': {
-        name: 'teamcity-bitbucket-visual-acceptance',
-        aliases: ['tbva'],
-        description: 'Run visual-acceptance based off Bitbucket message',
-        works: 'insideProject',
-        availableOptions: [{
-          name: 'image-directory',
-          type: String,
-          default: 'visual-acceptance',
-          description: 'The ember-cli-visual-acceptance directory where images are save'
-        }, {
-          name: 'branch',
-          type: String,
-          default: 'master',
-          description: 'branch to push to'
-        }, {
-          name: 'user',
-          type: String,
-          default: '',
-          description: 'Bitbucket username'
-        }, {
-          name: 'password',
-          type: String,
-          default: '',
-          description: 'Bitbucket user\'s password'
-        }, {
-          name: 'domain',
-          type: String,
-          default: '',
-          description: 'Domain of Bitbucket server ex("bitbucket.host.com")'
-        }, {
-          name: 'project',
-          type: String,
-          default: '',
-          description: 'Name of project where the repository is held'
-        }, {
-          name: 'repo',
-          type: String,
-          default: '',
-          description: 'Name of the repository'
-        }, {
-          name: 'api-url',
-          type: String,
-          default: '',
-          description: 'Url of api server to save and host images. https://gitlab.com/EWhite613/express-reports'
-        }, {
-          name: 'push-var',
-          type: String,
-          default: 'true',
-          description: 'Option that compares to pushOn to determine if it\'s okay to push'
-        },
-        {
-          name: 'push-on',
-          type: String,
-          default: 'true',
-          description: 'Option that compares to pushVar to determine if it\'s okay to push'
-        }],
-        run: function (options, rawArgs) {
-          if (options.user.length === 0 || options.password.length === 0 || options.domain.length === 0 ||
-           options.project.length === 0 || options.repo.length === 0 || options.apiUrl.length === 0) {
-            console.log('Need to supply a user, password, domain, project, repo, and api-url .' +
-             ' Sorry the bitbucket api sucks. \n Just running ember test')
-            return runCommand('ember', ['test'])
-          }
-
-          if (process.env.TEAMCITY_PULL_REQUEST === null) {
-            console.log('No Teamcity found. Just running ember test')
-            return runCommand('ember', ['test'])
-          }
-          process.env.NO_ASSERT = true
-          process.env.TEAMCITY_API_URL = options.apiUrl
-          var prNumber = process.env.TEAMCITY_PULL_REQUEST
-          if (prNumber !== false && prNumber !== 'false') {
-            return runCommand('ember', ['br']).then(function (params) {
-              return buildTeamcityBitbucketReport(params, options, prNumber)
-            }, function (params) {
-              return buildTeamcityBitbucketReport(params, options, prNumber).then(function (params) {
-                throw new Error('Exit 1')
-              })
-            })
-          } else if (prNumber === false || prNumber === 'false') {
-            return runCommand('ember', ['new-baseline', '--image-directory=' +
-             options.imageDirectory]).then(function (params) {
-               return runCommand('git', ['add', '-f', '-A', options.imageDirectory], true).then(function (params) {
-                 return runCommand('git', ['commit', '-m',
-                  'Adding new baseline images [ci skip]'], true).then(function (params) {
-                    if (options.pushVar === options.pushOn) {
-                      return runCommand('git', ['push', 'origin', 'HEAD:' + options.branch])
-                    }
-                  })
-               })
-             })
-          }
-        }
-      }
-    }
+    return require('./lib/commands')
   }
 }
