+++ lib/commands/index.js
@@ -1,0 +1,8 @@
+'use strict'
+
+module.exports = {
+  'new-baseline': require('./new-baseline.js'),
+  'build-report': require('./build-report.js'),
+  'teamcity-bitbucket-visual-acceptance': require('./teamcity-bitbucket-visual-acceptance.js'),
+  'travis-visual-acceptance': require('./travis-visual-acceptance.js')
+}
